@@ -26,17 +26,16 @@
         }
     ],
     "require": {
-        "php": "^7.2.0|^8.0",
-        "guzzlehttp/guzzle": "^6.3|^7.0",
-        "illuminate/notifications": "^5.6|^6.0|^7.0|^8.0|^9.0|^10.0|^11.0",
-        "illuminate/queue": "^5.6|^6.0|^7.0|^8.0|^9.0|^10.0|^11.0",
-        "illuminate/support": "^5.6|^6.0|^7.0|^8.0|^9.0|^10.0|^11.0"
+        "php": "^8.0",
+        "guzzlehttp/guzzle": "^6.3|^7.9.3",
+        "illuminate/notifications": "^10.0|^11.0|^12.0",
+        "illuminate/queue": "^10.0|^11.0|^12.0",
+        "illuminate/support": "^10.0|^11.0|^12.0"
     },
     "require-dev": {
-        "laravel/pint": "^1.22",
-        "mockery/mockery": "^1.3.3",
-        "pestphp/pest": "^3.8",
-        "pestphp/pest-plugin-drift": "^3.0"
+        "laravel/pint": "^1.22.1",
+        "mockery/mockery": "^1.6.12",
+        "pestphp/pest": "^3.8.2"
     },
     "autoload": {
         "psr-4": {
@@ -49,7 +48,8 @@
         }
     },
     "scripts": {
-        "test": "vendor/bin/phpunit"
+        "test": "vendor/bin/pest",
+        "lint": "vendor/bin/pint"
     },
     "config": {
         "sort-packages": true,
